backend/requirements.txt
# Enhanced VERSSAI Backend Requirements
# Python 3.8+ required

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# HTTP client for N8N integration
httpx==0.25.2

# File handling
aiofiles==23.2.1
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2

# Data processing
pandas==2.1.4
numpy==1.24.3

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.1

# Logging and monitoring
structlog==23.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# CORS handling
python-cors==1.7.0

# Async support
asyncio-mqtt==0.11.1

# Document parsing
python-magic==0.4.27
textract==1.6.5

# Background tasks
celery==5.3.4
redis==5.0.1
